refactor(gui): AppSignalHandler routes only -- extract session bootstrap + equipment helpers
Completes the "route, don't implement" audit (docs/todo/infra.md): every
subscribe lambda now takes the signal payload, calls one or two helpers,
and reflects results back into UI state.

- StartSessionSignal (255 lines -> 3 preconditions + one call): body moved
  verbatim to the new SessionBootstrapper.BuildAndStartAsync -- schedule
  build, config injection, session create, scout/guider/phase event
  wiring, tracked RunAsync + backlash mirror-back. Container-free (the
  lambda resolves ISessionFactory); verified by a line-multiset diff
  against the original.
- AssignDeviceSignal: target-switch -> EquipmentActions.ApplyAssignment
  (pure); orphan handling -> EquipmentActions.AutoDisconnectOrphanAsync
  returning (OrphanDisconnectOutcome, DisconnectSafety) that the lambda
  maps to notifications.
- ConnectDeviceSignal: resolve loop ->
  EquipmentActions.ResolveDeviceForConnect (hub-registered device
  preferred, discovery fallback by SameDevice).
- SetCoolerSetpointSignal / SetCoolerOffSignal ->
  EquipmentActions.SetCoolerSetpointAsync / SetCoolerOffAsync (immediate
  counterparts to the ramped WarmAndCoolerOffAsync).
- TakePreviewSignal: the remaining inline device-resolution block ->
  EquipmentActions.ResolveOtaCaptureDevices (capture itself already
  routed through LiveSessionActions.CaptureCameraPreviewAsync).
- infra.md checklist ticked, incl. two stale entries (SaveSnapshot /
  JogFocuser were already routed) and a reviewed-no-change note for the
  weather-refresh item; recorded a PlanetaryCaptureControllerTests
  full-suite timeout flake (passes in isolation) under Flaky Tests.

AppSignalHandler: 2,991 -> 2,720 lines. 0-warning build; full unit suite
3056/3057 green (the 1 failure is the recorded pre-existing planetary
timeout flake, unrelated to this change and passing in isolation);
functional suite 312/312.
