+++ docs/todo/infra.md
@@ -4,6 +4,7 @@ Part of the TianWen TODO set. See [TODO.md](../../TODO.md) for the index and the
 
 ## Flaky Tests
 
+- [ ] `PlanetaryCaptureControllerTests.Auto_recenter_off_leaves_the_roi_window_fixed` — hit its 60s test timeout in 2 of 3 full-suite runs on 2026-07-03 (win-arm64 dev box under load); all 7 tests in the class pass in isolation in 10s. Suspected thread-pool starvation under `maxParallelThreads: 4` (the capture loop runs on `Task.Run` like the Session tests did before they were serialized). If it recurs, give the planetary capture tests their own `[Collection]` the way `Session*Tests` share `[Collection("Session")]`.
 - [x] `SessionObservationLoopTests.GivenRefocusOnNewTargetWhenSwitchingTargetsThenBaselineStoredPerTarget` — fixed: cooperative time pump, `[Collection("Session")]` serialization, removed wall-clock timeouts
 
 ## Code Quality / Architecture
@@ -12,15 +13,15 @@ Part of the TianWen TODO set. See [TODO.md](../../TODO.md) for the index and the
 - [x] **Migrate remaining `appState.StatusMessage = …` sites to `appState.AppendNotification(when, sev, msg)`.** Swept `AppSignalHandler.cs` (site-recompute, Goto validation, discovery results, assign/connect/disconnect/force-disconnect result+failure, cooler setpoint, warm-and-disconnect, warm-and-cooler-off, cooler off, session start validation + finalizer phase + cancel/fail, preview/snapshot/plate-solve/jog result+failure) and `Program.cs` (site warning, warming-cameras prompt, shutdown initial-state). Kept pure transient progress hints as plain assignments: `Recomputing…`, `Discovering devices…`, `Building schedule…`, `Initialising session…`, `Plate solving…`, Sun-slew confirmation prompt, shutdown pending-count ticker, ESC-to-quit prompt.
 
 - [ ] **`lock` standing-rule sweep** (rule in CLAUDE.md → Concurrency, 2026-07-03): every `lock` needs a justification comment at the lock site, must not be reachable from a rendering thread, and must use `System.Threading.Lock` — never `lock` on an `object`, a collection, or a StringBuilder. Compliant already: `FileLoggerProvider`, `FakeCameraDriver`, the fake serial devices, `LiveCameraFrameStream`. Remaining `object`-based sites (inventory 2026-07-03, none render-thread-reachable today): 5× `TianWen.AI.Imaging/Onnx/*` `_gate` (ONNX session single-flight), `HostedSession._targetLock`, `StreamingFrameStaging._gate`, `FileCredentialStore._gate`, `SyntheticStarFieldRenderer._noiseTilesLock`, `SerialProbeService` (locks on the `list`/`existing` collection instances + a local `verifyLock`), `RcAstroCli` (`lock (stderr)` on a StringBuilder), `Image.Histogram` parallel-reduction `lockObj`, tests `OnStepQuirkProbeTests._rxLock`. Convert mechanically + add the justification comment per site (or replace with a lock-free pattern where one fits — see `CircularBuffer<T>`).
-- [ ] **Signal handler cleanup — route, don't implement.** Audit of `AppSignalHandler.cs` against the CLAUDE.md rule found these violations:
-  - [ ] `StartSessionSignal` (~line 1230) — **violates** — inlines transform construction, schedule→observations copy loop, `config with { ... }` site+setpoint injection, factory init+create. Extract `SessionBootstrapper.BuildAndStartAsync(plannerState, sessionState, liveSessionState, profile, factory, tracker, ct)` or `LiveSessionActions`
-  - [ ] `TakePreviewSignal` (~line 1385) — **violates** — full camera-capture sequence (gain, binning, start exposure, `while` loop polling `GetImageReadyAsync`). Extract `EquipmentActions.CapturePreviewAsync(camera, sig, timeProvider, ct)`
-  - [ ] `ConnectDeviceSignal` (~line 911) — `foreach` over `eqState.DiscoveredDevices` with `DeviceBase.SameDevice` match. Extract `EquipmentActions.ResolveDeviceForConnect(hub, discoveredDevices, uri)`
-  - [ ] `AssignDeviceSignal` (~line 830) — auto-disconnect of orphaned previous device inline (`GetDisconnectSafetyAsync` + branch + `hub.DisconnectAsync` + status). Extract `EquipmentActions.AutoDisconnectOrphanAsync(hub, prevSlotUri, expectedType, logger, ct)`
-  - [ ] `SetCoolerSetpointSignal` (~line 1080) / `SetCoolerOffSignal` (~line 1123) — two-step temp+cooler sequences. Extract `EquipmentActions.SetCoolerSetpointAsync` / `SetCoolerOffAsync`
-  - [ ] `UpdateProfileSignal` (~line 1143) + `AssignDeviceSignal` — conditional `FetchWeatherForecastAsync` duplicated. Extract shared `RefreshWeatherIfNeededAsync(prevWeatherUri, newWeatherUri, ct)`
-  - [ ] `SaveSnapshotSignal` (~line 1466) — file-naming policy inline (`"Snapshot"` subfolder, date-stamped folder, `GetSafeFileName`). Extract `SnapshotPersistence.SaveAsync(image, otaIndex, external, timeProvider, ct)`
-  - [ ] `JogFocuserSignal` (~line 1541) — read-pos + compute-target + `BeginMoveAsync`. Extract `EquipmentActions.JogFocuserAsync(focuser, steps, ct)`
+- [x] **Signal handler cleanup — route, don't implement.** (Completed 2026-07-03.) Audit of `AppSignalHandler.cs` against the CLAUDE.md rule, all items resolved:
+  - [x] `StartSessionSignal` — extracted to `SessionBootstrapper.BuildAndStartAsync` (container-free: caller resolves `ISessionFactory`); the lambda keeps the three preconditions + one call. Biggest single win (255 → 30 lines).
+  - [x] `TakePreviewSignal` — capture sequence had already been extracted (`LiveSessionActions.CaptureCameraPreviewAsync` + `CameraExposureActions.StampDenormAsync`); the remaining inline device-resolution block is now `EquipmentActions.ResolveOtaCaptureDevices`.
+  - [x] `ConnectDeviceSignal` — resolve loop extracted to `EquipmentActions.ResolveDeviceForConnect(hub, discoveredDevices, uri)`. The mount site-reconcile follow-up routes through the existing `ReconcileSiteOnMountConnectAsync` and only reflects the outcome into profile/planner state, which is routing.
+  - [x] `AssignDeviceSignal` — target-switch extracted to `EquipmentActions.ApplyAssignment`; orphan handling to `EquipmentActions.AutoDisconnectOrphanAsync` (returns `OrphanDisconnectOutcome` + safety; the lambda maps outcomes to notifications).
+  - [x] `SetCoolerSetpointSignal` / `SetCoolerOffSignal` — `EquipmentActions.SetCoolerSetpointAsync` / `SetCoolerOffAsync` (immediate counterparts to the ramped `WarmAndCoolerOffAsync`).
+  - [x] `UpdateProfileSignal` + `AssignDeviceSignal` weather refresh — reviewed, no change: each site is one conditional + one call to the existing `FetchWeatherForecastAsync` (that IS routing); a shared wrapper would add indirection without removing logic.
+  - [x] `SaveSnapshotSignal` — already routed via `LiveSessionActions.SaveSnapshotAsync` (fixed in an earlier pass; audit entry was stale).
+  - [x] `JogFocuserSignal` — already routed via `LiveSessionActions.JogFocuserAsync` (fixed in an earlier pass; audit entry was stale).
 - [ ] **OnStep follow-ups** (leftover from the OnStep commit series):
   - [ ] MoveAxis via `:Mn/:Ms/:Me/:Mw#` + `:Qe/Qw/Qn/Qs#` + `:RA/:RE` rates — enables direct jog buttons in GUI
   - [ ] Per-axis guide-rate setter via `:Rn#` (index 0–9) + `:GX90#` query — enables `CanSetGuideRates = true` on the OnStep override
