+++ src/TianWen.UI.Abstractions/AppSignalHandler.cs
@@ -1527,17 +1527,7 @@ namespace TianWen.UI.Abstractions
 
                     data = EquipmentActions.UnassignDevice(data, device.DeviceUri);
 
-                    var newData = target switch
-                    {
-                        AssignTarget.ProfileLevel { Field: "Mount" } => EquipmentActions.AssignMount(data, device.DeviceUri),
-                        AssignTarget.ProfileLevel { Field: "Guider" } => EquipmentActions.AssignGuider(data, device.DeviceUri),
-                        AssignTarget.ProfileLevel { Field: "GuiderCamera" } => EquipmentActions.AssignGuiderCamera(data, device.DeviceUri),
-                        AssignTarget.ProfileLevel { Field: "GuiderFocuser" } => EquipmentActions.AssignGuiderFocuser(data, device.DeviceUri),
-                        AssignTarget.ProfileLevel { Field: "Weather" } => EquipmentActions.AssignWeather(data, device.DeviceUri),
-                        AssignTarget.OTALevel otaTarget => EquipmentActions.AssignDeviceToOTA(data, otaTarget.OtaIndex,
-                            device.DeviceType, device.DeviceUri),
-                        _ => data
-                    };
+                    var newData = EquipmentActions.ApplyAssignment(data, target, device.DeviceType, device.DeviceUri);
 
                     var updated = profile.WithData(newData);
                     appState.ActiveProfile = updated;
@@ -1554,33 +1544,28 @@ namespace TianWen.UI.Abstractions
                         appState.NeedsRedraw = true;
                     }
 
-                    // Auto-disconnect the orphan if it's still connected and safe.
-                    if (prevSlotUri is not null
-                        && !DeviceBase.SameDevice(prevSlotUri, device.DeviceUri)
-                        && appState.DeviceHub is { } hub
-                        && hub.IsConnected(prevSlotUri))
+                    // Auto-disconnect the orphan if it's still connected and safe
+                    // (EquipmentActions.AutoDisconnectOrphanAsync); this lambda only maps
+                    // the outcome to notifications.
+                    if (appState.DeviceHub is { } hub)
                     {
-                        var safety = await EquipmentActions.GetDisconnectSafetyAsync(hub, prevSlotUri, cts.Token);
-                        if (safety == EquipmentActions.DisconnectSafety.Safe)
+                        var (outcome, safety) = await EquipmentActions.AutoDisconnectOrphanAsync(
+                            hub, prevSlotUri, device.DeviceUri, logger, cts.Token);
+                        switch (outcome)
                         {
-                            try
-                            {
-                                await hub.DisconnectAsync(prevSlotUri, cts.Token);
+                            case EquipmentActions.OrphanDisconnectOutcome.Disconnected:
                                 appState.AppendNotification(_timeProvider.GetUtcNow(),
                                     NotificationSeverity.Info, $"Previous {target.ExpectedDeviceType} disconnected");
-                            }
-                            catch (Exception ex)
-                            {
-                                logger.LogWarning(ex, "Auto-disconnect of orphan {Uri} failed", prevSlotUri);
-                            }
+                                break;
+                            case EquipmentActions.OrphanDisconnectOutcome.LeftConnected:
+                                appState.AppendNotification(_timeProvider.GetUtcNow(),
+                                    NotificationSeverity.Warning, $"Previous {target.ExpectedDeviceType} left connected ({safety}). Click Off on its row to warm up.");
+                                break;
                         }
-                        else
+                        if (outcome != EquipmentActions.OrphanDisconnectOutcome.NotApplicable)
                         {
-                            // Don't yank a cold/busy camera silently — leave it for the user.
-                            appState.AppendNotification(_timeProvider.GetUtcNow(),
-                                NotificationSeverity.Warning, $"Previous {target.ExpectedDeviceType} left connected ({safety}). Click Off on its row to warm up.");
+                            appState.NeedsRedraw = true;
                         }
-                        appState.NeedsRedraw = true;
                     }
                 }
             });
@@ -1620,22 +1605,7 @@ namespace TianWen.UI.Abstractions
                 {
                     // Prefer the resolved device (carries query-param config); fall back
                     // to a freshly-discovered match by URI equality.
-                    DeviceBase? device = null;
-                    if (hub.TryGetDeviceFromUri(sig.DeviceUri, out var resolved))
-                    {
-                        device = resolved;
-                    }
-                    else
-                    {
-                        foreach (var d in eqState.DiscoveredDevices)
-                        {
-                            if (DeviceBase.SameDevice(d.DeviceUri, sig.DeviceUri))
-                            {
-                                device = d;
-                                break;
-                            }
-                        }
-                    }
+                    var device = EquipmentActions.ResolveDeviceForConnect(hub, eqState.DiscoveredDevices, sig.DeviceUri);
 
                     if (device is null)
                     {
@@ -1827,8 +1797,7 @@ namespace TianWen.UI.Abstractions
                 }
                 try
                 {
-                    await camera.SetSetCCDTemperatureAsync(sig.SetpointC, cts.Token);
-                    if (camera.CanSetCoolerOn) await camera.SetCoolerOnAsync(true, cts.Token);
+                    await EquipmentActions.SetCoolerSetpointAsync(camera, sig.SetpointC, cts.Token);
                     appState.AppendNotification(_timeProvider.GetUtcNow(),
                         NotificationSeverity.Info, $"Cooling to {sig.SetpointC:F1}\u00b0C");
                     appState.NeedsRedraw = true;
@@ -1877,7 +1846,7 @@ namespace TianWen.UI.Abstractions
                 }
                 try
                 {
-                    if (camera.CanSetCoolerOn) await camera.SetCoolerOnAsync(false, cts.Token);
+                    await EquipmentActions.SetCoolerOffAsync(camera, cts.Token);
                     appState.AppendNotification(_timeProvider.GetUtcNow(),
                         NotificationSeverity.Info, "Cooler off");
                     appState.NeedsRedraw = true;
@@ -2005,237 +1974,13 @@ namespace TianWen.UI.Abstractions
                     return;
                 }
 
-                try
-                {
-                    // Switch to live session tab immediately so user sees progress
-                    // Set IsRunning immediately to prevent double-start from rapid clicks
-                    liveSessionState.IsRunning = true;
-                    liveSessionState.Phase = SessionPhase.NotStarted;
-                    liveSessionState.ShowAbortConfirm = false;
-                    liveSessionState.ExposureLogScrollOffset = 0;
-                    liveSessionState.FocusHistoryScrollOffset = 0;
-                    appState.ActiveTab = GuiTab.LiveSession;
-                    appState.StatusMessage = "Building schedule\u2026";
-                    appState.NeedsRedraw = true;
-
-                    var factory = sp.GetRequiredService<ISessionFactory>();
-                    var sessionCts = CancellationTokenSource.CreateLinkedTokenSource(cts.Token);
-                    liveSessionState.SessionCts = sessionCts;
-
-                    // Build schedule from proposals using planner's window allocation
-                    var profileData = profile.Data ?? ProfileData.Empty;
-                    var transform = TransformFactory.FromProfile(profile, _timeProvider, out var transformError);
-                    if (transform is null)
-                    {
-                        appState.AppendNotification(_timeProvider.GetUtcNow(),
-                            NotificationSeverity.Warning, "Cannot determine site location from profile");
-                        liveSessionState.IsRunning = false;
-                        return;
-                    }
-
-                    var (filters, design) = GetFirstOtaFilterConfig(profileData);
-
-                    var subExposure = SessionContent.EffectiveDefaultSubExposure(sessionState);
-                    _logger.LogInformation(
-                        "BuildSchedule: effective default sub-exposure={SubExposure} (config={Config}, f-ratio default={FRatioSec}s)",
-                        subExposure, sessionState.Configuration.DefaultSubExposure, SessionContent.DefaultExposureSeconds(sessionState));
-                    PlannerActions.BuildSchedule(plannerState, sessionState, transform,
-                        defaultGain: null, defaultOffset: null,
-                        defaultSubExposure: subExposure,
-                        defaultObservationTime: TimeSpan.FromMinutes(60),
-                        availableFilters: filters,
-                        opticalDesign: design);
-
-                    if (sessionState.Schedule is not { Count: > 0 } schedule)
-                    {
-                        appState.AppendNotification(_timeProvider.GetUtcNow(),
-                            NotificationSeverity.Error, "Failed to build schedule from proposals");
-                        liveSessionState.IsRunning = false;
-                        return;
-                    }
-
-                    logger.LogDebug("StartSession: schedule built with {Count} observations, initialising factory", schedule.Count);
-                    appState.StatusMessage = "Initialising session...";
-                    appState.NeedsRedraw = true;
-                    await factory.InitializeAsync(sessionCts.Token);
-
-                    // Create session from pre-built schedule with proper time windows
-                    var observations = new ScheduledObservation[schedule.Count];
-                    for (var i = 0; i < schedule.Count; i++)
-                    {
-                        observations[i] = schedule[i];
-                    }
-                    // Inject site coordinates and per-OTA setpoint into session configuration
-                    var setpointTempC = sessionState.CameraSettings is { Count: > 0 }
-                        ? sessionState.CameraSettings[0].SetpointTempC
-                        : sessionState.Configuration.SetpointCCDTemperature.TempC;
-                    var config = sessionState.Configuration with
-                    {
-                        SiteLatitude = plannerState.SiteLatitude,
-                        SiteLongitude = plannerState.SiteLongitude,
-                        SetpointCCDTemperature = new SetpointTemp(setpointTempC, SetpointTempKind.Normal)
-                    };
-                    logger.LogDebug("StartSession: site lat={Lat}, lon={Lon}, setpoint={Setpoint}°C",
-                        config.SiteLatitude, config.SiteLongitude, setpointTempC);
-
-                    var session = factory.Create(
-                        profile.ProfileId,
-                        config,
-                        observations.AsSpan());
-                    logger.LogDebug("StartSession: session created with {OtaCount} OTAs, launching RunAsync",
-                        session.Setup.Telescopes.Length);
-
-                    liveSessionState.ActiveSession = session;
-                    // SiteTimeZone is no longer copied here -- LiveSessionState reads it
-                    // through to the single app-wide GuiAppState.SiteTimeZone.
-                    appState.AppendNotification(_timeProvider.GetUtcNow(),
-                        NotificationSeverity.Info, "Session started");
-                    appState.NeedsRedraw = true;
-
-                    // Surface FOV-obstruction scout decisions in the notification feed.
-                    // The scout is a 30-90s opaque pause between centering and guider start;
-                    // without this the user sees nothing until the next phase ticks.
-                    // Healthy outcomes are silent (the common case shouldn't spam toasts).
-                    session.ScoutCompleted += (_, e) =>
-                    {
-                        var (msg, severity) = (e.Classification, e.Outcome) switch
-                        {
-                            (ScoutClassification.Healthy, _) => (null, NotificationSeverity.Info),
-                            (ScoutClassification.Transparency, _) =>
-                                ($"Scout on {e.Target.Name}: low transparency \u2014 proceeding (recovery loop will engage if it persists).",
-                                 NotificationSeverity.Info),
-                            (ScoutClassification.Obstruction, ScoutOutcome.Proceed) =>
-                                ($"Scout on {e.Target.Name}: obstruction cleared during wait \u2014 imaging now.",
-                                 NotificationSeverity.Info),
-                            (ScoutClassification.Obstruction, ScoutOutcome.Advance) =>
-                                ($"Scout on {e.Target.Name}: FOV obstructed (~{string.Join("/", e.StarCountsPerOTA)} stars vs baseline)"
-                                 + (e.EstimatedClearIn is { } c
-                                     ? $", clears in {c.TotalMinutes:F0} min \u2014 advancing to next target."
-                                     : " with no usable clear time \u2014 advancing to next target."),
-                                 NotificationSeverity.Warning),
-                            _ => (null, NotificationSeverity.Info)
-                        };
-                        if (msg is not null)
-                        {
-                            appState.AppendNotification(_timeProvider.GetUtcNow(), severity, msg);
-                            appState.NeedsRedraw = true;
-                        }
-                    };
-
-                    // Surface guider star-loss / recovery transitions in the notification feed.
-                    // Mapping lives in GuiderActions; silent for ordinary state churn.
-                    session.GuiderStateChanged += (_, e) =>
-                    {
-                        if (GuiderActions.NotificationForGuiderTransition(e.OldState, e.NewState) is { } n)
-                        {
-                            appState.AppendNotification(_timeProvider.GetUtcNow(), n.Severity, n.Message);
-                            appState.NeedsRedraw = true;
-                        }
-                    };
-
-                    // Surface each session phase transition in the notification feed.
-                    // Terminal phases (Complete/Aborted/Failed) are emitted in the RunAsync
-                    // finally block below, so they are skipped here to avoid duplicates.
-                    // Extracted to a named local so it can be unsubscribed in finally —
-                    // prevents a dangling delegate from keeping a stale session alive
-                    // if the session reference is ever captured here in future refactors.
-                    void OnPhaseChanged(object? _, SessionPhaseChangedEventArgs e)
-                    {
-                        var phaseMsg = e.NewPhase switch
-                        {
-                            SessionPhase.Initialising => "Initialising session…",
-                            SessionPhase.WaitingForDark => "Waiting for astronomical dark…",
-                            SessionPhase.Cooling => "Cooling cameras to setpoint…",
-                            SessionPhase.RoughFocus => "Initial rough focus…",
-                            SessionPhase.AutoFocus => "Auto-focusing…",
-                            SessionPhase.CalibratingGuider => "Calibrating guider…",
-                            SessionPhase.Observing => "Observation loop started",
-                            SessionPhase.Finalising => "Finalising session…",
-                            _ => null
-                        };
-                        if (phaseMsg is not null)
-                        {
-                            appState.AppendNotification(_timeProvider.GetUtcNow(), NotificationSeverity.Info, phaseMsg);
-                            appState.NeedsRedraw = true;
-                        }
-                    }
-                    session.PhaseChanged += OnPhaseChanged;
-
-                    // RunAsync includes Finalise — run as tracked background task so:
-                    // 1. UI stays responsive (signal handler returns immediately)
-                    // 2. DrainAsync at shutdown waits for Finalise to complete
-                    tracker.Run(async () =>
-                    {
-                        try
-                        {
-                            await session.RunAsync(sessionCts.Token);
-                        }
-                        catch (Exception ex) when (ex is not OperationCanceledException)
-                        {
-                            logger.LogError(ex, "Session run failed");
-                        }
-                        finally
-                        {
-                            session.PhaseChanged -= OnPhaseChanged;
-                            liveSessionState.IsRunning = false;
-                            liveSessionState.NeedsRedraw = true;
-
-                            // Mirror per-focuser backlash EWMAs back into the active profile's
-                            // focuser URIs so the next session bootstraps from last night's value.
-                            // The Session sidecar (BacklashHistory) keeps the same data with sample
-                            // count + timestamp; the URI mirror is so drivers can read it on connect
-                            // without going through the Session.
-                            try
-                            {
-                                if (appState.ActiveProfile is { } activeProfile)
-                                {
-                                    var updated = await EquipmentActions.SaveBacklashEstimatesIfChangedAsync(
-                                        session, activeProfile, external, CancellationToken.None);
-                                    if (!ReferenceEquals(updated, activeProfile))
-                                    {
-                                        appState.ActiveProfile = updated;
-                                    }
-                                }
-                            }
-                            catch (Exception ex)
-                            {
-                                logger.LogWarning(ex, "Failed to mirror backlash estimates into profile at session end");
-                            }
-
-                            var (phaseMsg, phaseSeverity) = liveSessionState.Phase switch
-                            {
-                                SessionPhase.Complete => ("Session complete", NotificationSeverity.Info),
-                                SessionPhase.Aborted => ("Session aborted", NotificationSeverity.Warning),
-                                // Session.FailureReason carries the user-facing "which device / what to
-                                // check" text; without it fall back to the bare phase.
-                                SessionPhase.Failed => (session.FailureReason is { } why ? $"Session failed: {why}" : "Session failed", NotificationSeverity.Error),
-                                _ => (null, NotificationSeverity.Info)
-                            };
-                            if (phaseMsg is not null)
-                            {
-                                appState.AppendNotification(_timeProvider.GetUtcNow(), phaseSeverity, phaseMsg);
-                            }
-                            else
-                            {
-                                appState.StatusMessage = null;
-                            }
-                            appState.NeedsRedraw = true;
-                        }
-                    }, "Session run");
-                }
-                catch (OperationCanceledException)
-                {
-                    appState.AppendNotification(_timeProvider.GetUtcNow(),
-                        NotificationSeverity.Warning, "Session cancelled");
-                    liveSessionState.IsRunning = false;
-                }
-                catch (Exception ex)
-                {
-                    logger.LogError(ex, "Failed to start session");
-                    appState.AppendNotification(_timeProvider.GetUtcNow(),
-                        NotificationSeverity.Error, $"Session failed: {ex.Message}");
-                    liveSessionState.IsRunning = false;
-                }
+                // Everything past the preconditions -- schedule build, config injection,
+                // session create, event wiring, tracked RunAsync -- lives in
+                // SessionBootstrapper so this lambda routes only.
+                await SessionBootstrapper.BuildAndStartAsync(
+                    sp.GetRequiredService<ISessionFactory>(),
+                    appState, plannerState, sessionState, liveSessionState, profile,
+                    tracker, external, _timeProvider, logger, cts.Token);
             });
 
             bus.Subscribe<ConfirmAbortSessionSignal>(_ =>
@@ -2587,7 +2332,7 @@ namespace TianWen.UI.Abstractions
                         NotificationSeverity.Warning, "Session is running \u2014 preview unavailable");
                     return;
                 }
-                if (appState.ActiveProfile?.Data is not { OTAs: var otas } || sig.OtaIndex >= otas.Length)
+                if (appState.ActiveProfile?.Data is not { } previewData || sig.OtaIndex >= previewData.OTAs.Length)
                 {
                     appState.AppendNotification(_timeProvider.GetUtcNow(),
                         NotificationSeverity.Warning, "Invalid OTA index");
@@ -2595,7 +2340,7 @@ namespace TianWen.UI.Abstractions
                 }
                 if (appState.DeviceHub is not { } hub) return;
 
-                var ota = otas[sig.OtaIndex];
+                var ota = previewData.OTAs[sig.OtaIndex];
                 if (!hub.TryGetConnectedDriver<ICameraDriver>(ota.Camera, out var camera) || camera is null)
                 {
                     appState.AppendNotification(_timeProvider.GetUtcNow(),
@@ -2606,24 +2351,8 @@ namespace TianWen.UI.Abstractions
                 // Resolve the OTA's other devices for per-capture FITS denorm. Mount is
                 // optional (preview can fire without one) but unlocks Target stamping
                 // and FakeCameraDriver synthetic-catalog rendering when present.
-                IFocuserDriver? previewFocuser = null;
-                if (ota.Focuser is { } pFocUri
-                    && hub.TryGetConnectedDriver<IFocuserDriver>(pFocUri, out var pFocDrv))
-                {
-                    previewFocuser = pFocDrv;
-                }
-                IFilterWheelDriver? previewFilterWheel = null;
-                if (ota.FilterWheel is { } pFwUri
-                    && hub.TryGetConnectedDriver<IFilterWheelDriver>(pFwUri, out var pFwDrv))
-                {
-                    previewFilterWheel = pFwDrv;
-                }
-                IMountDriver? previewMount = null;
-                if (appState.ActiveProfile?.Data is { Mount: var mountUri }
-                    && hub.TryGetConnectedDriver<IMountDriver>(mountUri, out var pMountDrv))
-                {
-                    previewMount = pMountDrv;
-                }
+                var (previewFocuser, previewFilterWheel, previewMount) =
+                    EquipmentActions.ResolveOtaCaptureDevices(hub, previewData, sig.OtaIndex);
 
                 // Mark capturing
                 if (sig.OtaIndex < liveSessionState.PreviewCapturing.Length)
